dotbot/swarm: rename the sdk subpackage to swarm
The whole pydotbot package is branded the DotBot SDK, so an sdk
subpackage inside it was redundant. swarm names the thing the API
drives and deliberately pairs with the dotbot swarm CLI namespace:
one fleet, operated from the CLI, driven from Python.

AI-assisted: Claude Fable 5
