+++ dotbot/examples/sdk/pulse.py
@@ -6,7 +6,7 @@ python pulse.py [--swarm-url http://localhost:8000]
 import asyncio
 import math
 
-from dotbot.sdk import Swarm
+from dotbot.swarm import Swarm
 
 CENTER_X, CENTER_Y = 1000, 1000
